+++ lib/moneys.dart
@@ -2,9 +2,9 @@ library moneys;
 
 export 'src/enumerations/currency.dart';
 export 'src/enumerations/income_or_expense.dart';
-export 'src/enumerations/payment_method.dart';
 export 'src/enumerations/renewal.dart';
+export 'src/enumerations/transaction_method.dart';
 export 'src/models/budget.dart';
 export 'src/models/money.dart';
-export 'src/models/payment.dart';
-export 'src/models/payments.dart';
+export 'src/models/transaction.dart';
+export 'src/models/transactions.dart';